harness: absorb policy into substrate
