+++ spark/tests/test_recurrent.py
@@ -27,7 +27,7 @@ THIS = Path(__file__).resolve()
 SPARK_DIR = THIS.parent.parent
 sys.path.insert(0, str(SPARK_DIR))
 
-from harness.policy import default_policy, RoleConfig  # noqa: E402
+from harness.substrate import default_policy, RoleConfig  # noqa: E402
 from harness.providers import (  # noqa: E402
     NormalizedResponse,
     ProviderRegistry,
